+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [8.2.2] - Unreleased
 
 - Fix Project Creation UI issue
+- Fix issue with whitespace confusing encryption key paths
 
 ## [8.2.1] - 2024-07-18